requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
selenium>=4.15.2
webdriver-manager>=4.0.1
supabase>=2.3.0
transformers>=4.35.2
torch>=2.1.0
Pillow>=10.1.0
numpy>=1.26.0
pandas>=2.1.0
python-dotenv>=1.0.0
